try 2

+++ app.py
@@ -663,7 +663,7 @@ with tab_movement:
             "Bet?": False,
         })
         for numcol in ["Proj", "Edge", f"First {line_word}", f"Latest {line_word}", "Move"]:
-            grid[numcol] = pd.to_numeric(grid[numcol], errors="coerce")
+            grid[numcol] = pd.to_numeric(grid[numcol], errors="coerce").astype("float64")
 
         edited = st.data_editor(
             grid, width='stretch', hide_index=True,
